extract ir_plan.cpp
